Move iiif.legacyinclusions to iiif.presentation.v2
