+++ src/Wellcome.Dds/IIIF/Auth/AuthCookieService1.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using IIIF.LegacyInclusions;
+using IIIF.Presentation.V2;
 using Newtonsoft.Json;
 
 namespace IIIF.Auth
