+++ src/Wellcome.Dds/IIIF/ImageApi/Service/ImageService2.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using IIIF.LegacyInclusions;
+using IIIF.Presentation.V2;
 using Newtonsoft.Json;
 
 namespace IIIF.ImageApi.Service
