+++ src/Wellcome.Dds/IIIF/Presentation/V2/AnnotationList.cs
@@ -2,7 +2,7 @@
 using IIIF.Presentation.V3.Annotation;
 using Newtonsoft.Json;
 
-namespace IIIF.LegacyInclusions
+namespace IIIF.Presentation.V2
 {
     public class AnnotationList : LegacyResourceBase
     {
