+++ src/Wellcome.Dds/IIIF/Search/V1/SearchResultsLayer.cs
@@ -1,4 +1,4 @@
-﻿using IIIF.LegacyInclusions;
+﻿using IIIF.Presentation.V2;
 using Newtonsoft.Json;
 
 namespace IIIF.Search.V1
