+++ src/Wellcome.Dds/Wellcome.Dds.Repositories/Presentation/IIIFBuilder.cs
@@ -4,8 +4,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using DlcsWebClient.Config;
 using IIIF;
-using IIIF.LegacyInclusions;
 using IIIF.Presentation;
+using IIIF.Presentation.V2;
 using IIIF.Presentation.V3;
 using IIIF.Presentation.V3.Annotation;
 using IIIF.Presentation.V3.Constants;
@@ -521,7 +521,7 @@ namespace Wellcome.Dds.Repositories.Presentation
             };
         }
 
-        private IIIF.LegacyInclusions.Annotation GetOATextLineAnnotation(
+        private IIIF.Presentation.V2.Annotation GetOATextLineAnnotation(
             AnnotationPage altoPage, TextLine tl, int lineIndex, string canvasId)
         {
             return new()
@@ -538,10 +538,10 @@ namespace Wellcome.Dds.Repositories.Presentation
             };
         }
 
-        private IIIF.LegacyInclusions.Annotation GetOAIllustrationAnnotation(
+        private IIIF.Presentation.V2.Annotation GetOAIllustrationAnnotation(
             AnnotationPage altoPage, Illustration il, int index, string canvasId)
         {
-            return new IIIF.LegacyInclusions.Annotation
+            return new IIIF.Presentation.V2.Annotation
             {
                 Id = uriPatterns.CanvasClassifyingAnnotation(
                     altoPage.ManifestationIdentifier, altoPage.AssetIdentifier, $"i{index}"),
@@ -746,7 +746,7 @@ namespace Wellcome.Dds.Repositories.Presentation
             
             foreach (var jItem in v3["items"])
             {
-                var annotation = new IIIF.LegacyInclusions.Annotation
+                var annotation = new IIIF.Presentation.V2.Annotation
                 {
                     Id = jItem.Value<string>("id"),
                     On = jItem["target"]?.Value<string>("id") ?? string.Empty
