+++ src/Wellcome.Dds/Wellcome.Dds/IIIFBuilding/IIIIFBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using IIIF;
-using IIIF.LegacyInclusions;
+using IIIF.Presentation.V2;
 using IIIF.Search.V1;
 using Newtonsoft.Json.Linq;
 using Wellcome.Dds.AssetDomain.Mets;
